requirements.txt
PyYAML==3.11
import-file==1.11
requests==2.7.0
six==1.9.0
slackclient==0.15
websocket-client==0.32.0
